user_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.27
asyncpg==0.29.0
pydantic==2.5.0
python-dotenv==1.0.0
bcrypt==4.1.2
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
